Show default selinux context as untrusted_app
In the CONFIG_FILE, there is a field for selinux context label.
Though this field is not wildely used by APatch or Android modules, it is confusing to have its default always to be `magisk`.

For example, if one sets an app to be excluded, then in the file `/data/adb/ap/package_config`, its selinux context is shown to be `u:r:magisk:s0`, which is not reasonable.
It is better to use the Android default context for applications to avoid confusions (of module developers who parse this file).